+++ .build.ps1
@@ -146,8 +146,8 @@ task Test {
 	$output = Invoke-Build . Demo\.build.ps1 -Result result | Out-String -Width:9999
 	if ($SkipTestDiff) { return }
 
-	assert (177 -eq $result.Tasks.Count) $result.Tasks.Count
-	assert (37 -eq $result.Errors.Count) $result.Errors.Count
+	assert (182 -eq $result.Tasks.Count) $result.Tasks.Count
+	assert (38 -eq $result.Errors.Count) $result.Errors.Count
 	assert ($result.Warnings.Count -ge 1)
 
 	# process and save the output