+++ Demo/Use.test.ps1
@@ -22,16 +22,44 @@ use ([System.Runtime.InteropServices.RuntimeEnvironment]::GetRuntimeDirectory())
 # It is fine to change the location (used for -If) and leave it changed.
 Set-Location "$env:windir\Microsoft.NET\Framework"
 
-# This task calls MSBuild 2.0 and tests its version.
-task v2.0.50727 -If (Test-Path 'v2.0.50727') {
+# These tasks calls MSBuild X.Y and test its version.
+
+task Version.2.0 -If (Test-Path 'HKLM:\SOFTWARE\Microsoft\MSBuild\ToolsVersions\2.0') {
+	use 2.0 MSBuild
+	$version = exec { MSBuild /version /nologo }
+	$version
+	assert ($version -like '2.0.*')
+}
+
+task Framework.2.0.50727 -If (Test-Path 'v2.0.50727') {
 	use Framework\v2.0.50727 MSBuild
 	$version = exec { MSBuild /version /nologo }
 	$version
 	assert ($version -like '2.0.*')
 }
 
-# This task calls MSBuild 4.0 and tests its version.
-task v4.0.30319 -If (Test-Path 'v4.0.30319') {
+task Version.3.5 -If (Test-Path 'HKLM:\SOFTWARE\Microsoft\MSBuild\ToolsVersions\3.5') {
+	use 3.5 MSBuild
+	$version = exec { MSBuild /version /nologo }
+	$version
+	assert ($version -like '3.5.*')
+}
+
+task Framework.3.5 -If (Test-Path 'v3.5') {
+	use Framework\v3.5 MSBuild
+	$version = exec { MSBuild /version /nologo }
+	$version
+	assert ($version -like '3.5.*')
+}
+
+task Version.4.0 -If (Test-Path 'HKLM:\SOFTWARE\Microsoft\MSBuild\ToolsVersions\4.0') {
+	use 4.0 MSBuild
+	$version = exec { MSBuild /version /nologo }
+	$version
+	assert ($version -like '4.0.*')
+}
+
+task Framework.4.0.30319 -If (Test-Path 'v4.0.30319') {
 	use Framework\v4.0.30319 MSBuild
 	$version = exec { MSBuild /version /nologo }
 	$version
@@ -51,9 +79,9 @@ task ResolvedPath {
 	assert (($path -like '?:\*\MyTestAlias') -or ($path -like '\\*\MyTestAlias'))
 }
 
-# Error: missing framework.
-task MissingFramework {
-	use Framework\MissingFramework MSBuild
+# Error: missing version.
+task MissingVersion {
+	use 3.14 MSBuild
 }
 
 # Error: invalid framework.
@@ -61,9 +89,9 @@ task InvalidFramework {
 	use 'Framework\<>' MSBuild
 }
 
-# Error: missing directory.
-task MissingDirectory {
-	use MissingDirectory MyScript
+# Error: missing framework.
+task MissingFramework {
+	use Framework\MissingFramework MSBuild
 }
 
 # Error: invalid directory.
@@ -71,20 +99,23 @@ task InvalidDirectory {
 	use '\<>' MyScript
 }
 
+# Error: missing directory.
+task MissingDirectory {
+	use MissingDirectory MyScript
+}
+
 # The default task calls the others and checks that InvalidFramework and
 # DoNotDotSource have failed. Failing tasks are referenced as @{Task=1}.
 task . `
-v2.0.50727,
-v4.0.30319,
-CurrentFramework,
-ResolvedPath,
+@{MissingVersion=1},
 @{MissingFramework=1},
 @{InvalidFramework=1},
 @{MissingDirectory=1},
 @{InvalidDirectory=1},
 {
-	Test-Error MissingFramework "Missing directory '*\Microsoft.NET\Framework\MissingFramework'.*Framework\MissingFramework MSBuild*"
-	Test-Error InvalidFramework "Missing directory '*\Microsoft.NET\Framework\<>'.*'Framework\<>'*"
-	Test-Error MissingDirectory "Missing directory '*\MissingDirectory'.*MissingDirectory MyScript*"
-	Test-Error InvalidDirectory "*'\<>' MyScript*"
+	Test-Error MissingVersion   "Cannot resolve '3.14'.*"
+	Test-Error InvalidFramework "Cannot resolve 'Framework\<>'.*"
+	Test-Error MissingFramework "Cannot resolve 'Framework\MissingFramework'.*"
+	Test-Error InvalidDirectory "Cannot resolve '\<>'.*"
+	Test-Error MissingDirectory "Cannot resolve 'MissingDirectory'.*"
 }